Changed Editing Profile
- removed EditProfileFragment
- added Fragments for editing of each name, institution and password
