+++ .idea/misc.xml
@@ -34,11 +34,14 @@
         <entry key="..\:/Users/Phili/Desktop/ase-ws-21-unser-horsaal/app/src/main/res/layout/fragment_create_course.xml" value="0.12158054711246201" />
         <entry key="..\:/Users/Phili/Desktop/ase-ws-21-unser-horsaal/app/src/main/res/layout/fragment_current_course.xml" value="0.264" />
         <entry key="..\:/Users/Phili/Desktop/ase-ws-21-unser-horsaal/app/src/main/res/layout/fragment_edit_profile.xml" value="0.20606884057971014" />
+        <entry key="..\:/Users/Phili/Desktop/ase-ws-21-unser-horsaal/app/src/main/res/layout/fragment_edit_profile_institution.xml" value="0.27291666666666664" />
+        <entry key="..\:/Users/Phili/Desktop/ase-ws-21-unser-horsaal/app/src/main/res/layout/fragment_edit_profile_name.xml" value="0.2869791666666667" />
+        <entry key="..\:/Users/Phili/Desktop/ase-ws-21-unser-horsaal/app/src/main/res/layout/fragment_edit_profile_password.xml" value="0.3625" />
         <entry key="..\:/Users/Phili/Desktop/ase-ws-21-unser-horsaal/app/src/main/res/layout/fragment_enter_course.xml" value="0.13405797101449277" />
         <entry key="..\:/Users/Phili/Desktop/ase-ws-21-unser-horsaal/app/src/main/res/layout/fragment_enter_or_create_course.xml" value="0.13405797101449277" />
         <entry key="..\:/Users/Phili/Desktop/ase-ws-21-unser-horsaal/app/src/main/res/layout/fragment_login.xml" value="0.264" />
         <entry key="..\:/Users/Phili/Desktop/ase-ws-21-unser-horsaal/app/src/main/res/layout/fragment_placeholder.xml" value="0.140625" />
-        <entry key="..\:/Users/Phili/Desktop/ase-ws-21-unser-horsaal/app/src/main/res/layout/fragment_profile.xml" value="0.2" />
+        <entry key="..\:/Users/Phili/Desktop/ase-ws-21-unser-horsaal/app/src/main/res/layout/fragment_profile.xml" value="0.264" />
         <entry key="..\:/Users/Phili/Desktop/ase-ws-21-unser-horsaal/app/src/main/res/layout/fragment_registration.xml" value="0.264" />
         <entry key="..\:/Users/Phili/Desktop/ase-ws-21-unser-horsaal/app/src/main/res/layout/participant_list_item.xml" value="0.264" />
         <entry key="..\:/Users/Phili/Desktop/ase-ws-21-unser-horsaal/app/src/main/res/layout/simple_course_item.xml" value="0.16799999999999998" />
@@ -61,19 +64,7 @@
         <entry key="..\:/Users/Ruslan/Documents/GitHub/ase-ws-21-unser-horsaal/app/src/main/res/layout/fragment_registration.xml" value="0.23324275362318841" />
         <entry key="..\:/Users/julia/Documents/GitHub/ase-ws-21-unser-horsaal/app/src/main/res/layout/activity_main.xml" value="0.1" />
         <entry key="..\:/Users/julia/Documents/GitHub/ase-ws-21-unser-horsaal/app/src/main/res/layout/activity_second.xml" value="0.1" />
-        <entry key="..\:/Users/julia/Documents/GitHub/ase-ws-21-unser-horsaal/app/src/main/res/layout/fragment_courses.xml" value="0.1" />
-        <entry key="..\:/Users/julia/Documents/GitHub/ase-ws-21-unser-horsaal/app/src/main/res/layout/fragment_profile.xml" value="0.15989583333333332" />
         <entry key="app/src/main/res/layout/activity_main.xml" value="0.21875" />
-        <entry key="app/src/main/res/layout/fragment_login.xml" value="0.18" />
-        <entry key="app/src/main/res/layout/fragment_profile.xml" value="0.1" />
-        <entry key="app/src/main/res/layout/fragment_registration.xml" value="0.2128623188405797" />
-<<<<<<< HEAD
-=======
-        <entry key="app/src/main/res/menu/bottom_navigation_menu.xml" value="0.22395833333333334" />
-        <entry key="app/src/main/res/menu/course_history_fragment_toolbar.xml" value="0.22395833333333334" />
-        <entry key="app/src/main/res/menu/edit_profile_fragment_toolbar.xml" value="0.22395833333333334" />
-        <entry key="app/src/main/res/menu/profile_fragment_toolbar.xml" value="0.22395833333333334" />
->>>>>>> 20b2eda (databinding for profile fragment)
       </map>
     </option>
   </component>
