+++ app/src/main/res/anim/slide_in_left.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<set xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shareInterpolator="false" >
+    <translate android:duration="200" android:fromXDelta="-100%" android:toXDelta="0%"/>
+    <alpha android:duration="200" android:fromAlpha="0.0" android:toAlpha="1.0" />
+</set>
