+++ app/src/main/res/anim/slide_out_left.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<set xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shareInterpolator="false" >
+    <translate android:duration="200" android:fromXDelta="0%" android:toXDelta="-100%"/>
+    <alpha android:duration="200" android:fromAlpha="1.0" android:toAlpha="0.0" />
+</set>
