+++ app/src/main/res/navigation/nav_graph.xml
@@ -88,9 +88,26 @@
             app:popUpTo="@id/coursesFragment"
             app:popUpToInclusive="true" />
         <action
-            android:id="@+id/action_profileFragment_to_editProfileFragment"
-            app:destination="@id/editProfileFragment"
-            app:popUpToInclusive="false" />
+            android:id="@+id/action_profileFragment_to_editProfileNameFragment"
+            app:destination="@id/editProfileNameFragment"
+            app:enterAnim="@anim/slide_in_right"
+            app:exitAnim="@anim/slide_out_left"
+            app:popEnterAnim="@anim/slide_in_left"
+            app:popExitAnim="@anim/slide_out_right" />
+        <action
+            android:id="@+id/action_profileFragment_to_editProfileInstitutionFragment"
+            app:destination="@id/editProfileInstitutionFragment"
+            app:enterAnim="@anim/slide_in_right"
+            app:exitAnim="@anim/slide_out_left"
+            app:popEnterAnim="@anim/slide_in_left"
+            app:popExitAnim="@anim/slide_out_right" />
+        <action
+            android:id="@+id/action_profileFragment_to_editProfilePasswordFragment"
+            app:destination="@id/editProfilePasswordFragment"
+            app:enterAnim="@anim/slide_in_right"
+            app:exitAnim="@anim/slide_out_left"
+            app:popEnterAnim="@anim/slide_in_left"
+            app:popExitAnim="@anim/slide_out_right" />
     </fragment>
     <fragment
         android:id="@+id/coursesFragment"
@@ -112,17 +129,6 @@
         <action
             android:id="@+id/action_coursesFragment_to_courseHistoryFragment"
             app:destination="@id/courseHistoryFragment" />
-    </fragment>
-    <fragment
-        android:id="@+id/editProfileFragment"
-        android:name="com.example.unserhoersaal.views.EditProfileFragment"
-        android:label="fragment_edit_profile"
-        tools:layout="@layout/fragment_edit_profile" >
-        <action
-            android:id="@+id/action_editProfileFragment_to_profileFragment"
-            app:destination="@id/profileFragment"
-            app:popUpTo="@id/profileFragment"
-            app:popUpToInclusive="true" />
     </fragment>
     <fragment
         android:id="@+id/courseHistoryFragment"
@@ -168,4 +174,49 @@
         android:name="com.example.unserhoersaal.views.CourseMeetingFragment"
         android:label="fragment_course_meeting"
         tools:layout="@layout/fragment_course_meeting" />
+    <fragment
+        android:id="@+id/editProfileNameFragment"
+        android:name="com.example.unserhoersaal.views.EditProfileNameFragment"
+        android:label="EditProfileNameFragment"
+        tools:layout="@layout/fragment_edit_profile_name">
+        <action
+            android:id="@+id/action_editProfileNameFragment_to_profileFragment"
+            app:destination="@id/profileFragment"
+            app:enterAnim="@anim/slide_in_left"
+            app:exitAnim="@anim/slide_out_right"
+            app:popEnterAnim="@anim/slide_in_right"
+            app:popExitAnim="@anim/slide_out_left"
+            app:popUpTo="@id/profileFragment"
+            app:popUpToInclusive="true" />
+    </fragment>
+    <fragment
+        android:id="@+id/editProfileInstitutionFragment"
+        android:name="com.example.unserhoersaal.views.EditProfileInstitutionFragment"
+        android:label="fragment_edit_profile_institution"
+        tools:layout="@layout/fragment_edit_profile_institution" >
+        <action
+            android:id="@+id/action_editProfileInstitutionFragment_to_profileFragment"
+            app:destination="@id/profileFragment"
+            app:enterAnim="@anim/slide_in_left"
+            app:exitAnim="@anim/slide_out_right"
+            app:popEnterAnim="@anim/slide_in_right"
+            app:popExitAnim="@anim/slide_out_left"
+            app:popUpTo="@id/profileFragment"
+            app:popUpToInclusive="true" />
+    </fragment>
+    <fragment
+        android:id="@+id/editProfilePasswordFragment"
+        android:name="com.example.unserhoersaal.views.EditProfilePasswordFragment"
+        android:label="fragment_edit_profile_password"
+        tools:layout="@layout/fragment_edit_profile_password" >
+        <action
+            android:id="@+id/action_editProfilePasswordFragment_to_profileFragment"
+            app:destination="@id/profileFragment"
+            app:enterAnim="@anim/slide_in_left"
+            app:exitAnim="@anim/slide_out_right"
+            app:popEnterAnim="@anim/slide_in_right"
+            app:popExitAnim="@anim/slide_out_left"
+            app:popUpTo="@id/profileFragment"
+            app:popUpToInclusive="true" />
+    </fragment>
 </navigation>
