+++ app/src/main/res/values/strings.xml
@@ -110,4 +110,5 @@
     <string name="dialog_delete_account_message">Möchten Sie Ihren Account wirklich löschen?</string>
     <string name="dialog_delete_account_true">OK</string>
     <string name="dialog_delete_account_false">Abbrechen</string>
+    <string name="edit_profile_name_hint">Nutzername</string>
 </resources>
